refactor: move fix_tectonic_links to TectonicUnit file
